requirements.txt
paho-mqtt
python-dotenv
pytz
requests